chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/at-scroll-timeline-start-end-expected.txt

This is a testharness.js-based test.
PASS Scroll at offset 0 updates animation correctly [element_0px_100px]
FAIL Scroll at offset 1 updates animation correctly [element_0px_100px] assert_equals: expected "101px" but got "100px"
FAIL Scroll at offset 50 updates animation correctly [element_0px_100px] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 99 updates animation correctly [element_0px_100px] assert_equals: expected "199px" but got "100px"
FAIL Scroll at offset 100 updates animation correctly [element_0px_100px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 101 updates animation correctly [element_0px_100px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_50px_100px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 1 updates animation correctly [element_50px_100px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 49 updates animation correctly [element_50px_100px] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 50 updates animation correctly [element_50px_100px]
FAIL Scroll at offset 51 updates animation correctly [element_50px_100px] assert_equals: expected "102px" but got "100px"
FAIL Scroll at offset 99 updates animation correctly [element_50px_100px] assert_equals: expected "198px" but got "100px"
FAIL Scroll at offset 100 updates animation correctly [element_50px_100px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 101 updates animation correctly [element_50px_100px] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 0 updates animation correctly [element_auto_auto]
FAIL Scroll at offset 21 updates animation correctly [element_auto_auto] assert_approx_equals: expected 110 +/- 5 but got 100
FAIL Scroll at offset 107 updates animation correctly [element_auto_auto] assert_approx_equals: expected 150 +/- 5 but got 100
FAIL Scroll at offset 193 updates animation correctly [element_auto_auto] assert_approx_equals: expected 190 +/- 5 but got 100
FAIL Scroll at offset 215 updates animation correctly [element_auto_auto] assert_approx_equals: expected 0 +/- 5 but got 100
FAIL Scroll at offset 0 updates animation correctly [element_50px_auto] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 49 updates animation correctly [element_50px_auto] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 50 updates animation correctly [element_50px_auto]
FAIL Scroll at offset 132 updates animation correctly [element_50px_auto] assert_approx_equals: expected 150 +/- 5 but got 100
FAIL Scroll at offset 198 updates animation correctly [element_50px_auto] assert_approx_equals: expected 190 +/- 5 but got 100
FAIL Scroll at offset 215 updates animation correctly [element_50px_auto] assert_approx_equals: expected 0 +/- 5 but got 100
PASS Scroll at offset 0 updates animation correctly [element_auto_100px]
FAIL Scroll at offset 1 updates animation correctly [element_auto_100px] assert_equals: expected "101px" but got "100px"
FAIL Scroll at offset 50 updates animation correctly [element_auto_100px] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 99 updates animation correctly [element_auto_100px] assert_equals: expected "199px" but got "100px"
FAIL Scroll at offset 100 updates animation correctly [element_auto_100px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_25p_75p] assert_approx_equals: expected 0 +/- 5 but got 100
FAIL Scroll at offset 21 updates animation correctly [element_25p_75p] assert_approx_equals: expected 0 +/- 5 but got 100
FAIL Scroll at offset 43 updates animation correctly [element_25p_75p] assert_approx_equals: expected 0 +/- 5 but got 100
FAIL Scroll at offset 75 updates animation correctly [element_25p_75p] assert_approx_equals: expected 120 +/- 5 but got 100
FAIL Scroll at offset 118 updates animation correctly [element_25p_75p] assert_approx_equals: expected 160 +/- 5 but got 100
FAIL Scroll at offset 172 updates animation correctly [element_25p_75p] assert_approx_equals: expected 0 +/- 5 but got 100
FAIL Scroll at offset 215 updates animation correctly [element_25p_75p] assert_approx_equals: expected 0 +/- 5 but got 100
FAIL Scroll at offset 63 updates animation correctly [element_calc_calc] assert_approx_equals: expected 0 +/- 5 but got 100
FAIL Scroll at offset 117 updates animation correctly [element_calc_calc] assert_approx_equals: expected 150 +/- 5 but got 100
FAIL Scroll at offset 182 updates animation correctly [element_calc_calc] assert_approx_equals: expected 0 +/- 5 but got 100
Harness: the test ran to completion.
